Extract query zero-result exit helpers
